examples/example_1_poetry/pyproject.toml
[tool.poetry]
name = "example-1-poetry"
version = "0.1.0"
description = ""
authors = ["tiebingice <1206345694@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
typing_inspect = "^0.9.0"
fastapi-utils = "^0.8.0"
pydantic_settings = "^2.7.1"
jinja2 = "^3.1.5"
tortoise-orm = "^0.23.0"
uvicorn = "^0.34.0"
asyncmy = "^0.2.10"


[tool.poetry.group.dev.dependencies]
fastapi-app-manage = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
